pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zrc"
version = "1.1.0"
description = "Zenoh Robot Control - A communication library for robotics based on Zenoh"
readme = "README.md"
authors = [{name = "ZRC Contributors"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]
requires-python = ">=3.7"
dependencies = [
    "eclipse-zenoh>=1.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/your-org/zrc"
Documentation = "https://github.com/your-org/zrc/blob/main/doc.md"
Repository = "https://github.com/your-org/zrc"